conflicted: An Alternative Conflict Resolution Strategy

R's default conflict management system gives the most recently loaded package precedence. This can make it hard to detect conflicts, particularly when they arise because a package update creates ambiguity that did not previously exist. 'conflicted' takes a different approach, making every conflict an error and forcing you to choose which function to use.

Resolving conflicts

To permanently resolve a conflict within a session, use assignment:

library(conflicted)
library(dplyr)

filter <- dplyr::filter

See also

Author

Maintainer: Hadley Wickham hadley@rstudio.com

Other contributors:

RStudio [copyright holder]