Persistently prefer one function over another

conflict_prefer() allows you to declare "winners" of conflicts. You can either declare a specific pairing (i.e. dplyr::filter() beats base::filter()), or an overall winner (i.e. dplyr::filter() beats all comers).

conflict_prefer(name, winner, losers = NULL, quiet = FALSE)

Arguments

name
Name of function.

winner
Name of package that should win the conflict.

losers
Optional vector of packages that should lose the conflict. If omitted, winner will beat all comers.

quiet
If TRUE, all output will be suppressed

Best practices

I recommend placing calls to conflict_prefer() at the top of your script, immediately underneath the relevant library() call.

Examples

# Prefer over all other packages
conflict_prefer("filter", "dplyr")
#> [conflicted] Will prefer dplyr::filter over any other package

# Prefer over specified package or packages
conflict_prefer("filter", "dplyr", "base")
#> [conflicted] Removing existing preference
#> [conflicted] Will prefer dplyr::filter over base::filter
conflict_prefer("filter", "dplyr", c("base", "filtration"))
#> [conflicted] Removing existing preference
#> [conflicted] Will prefer dplyr::filter over base::filter, filtration::filter