conflict_scout {conflicted}	R Documentation

Find conflicts amongst a set of packages

Description

conflict_scout() is the workhorse behind the conflicted package. You can call it directly yourself if you want to see all conflicts before hitting them in practice.

Usage

conflict_scout(pkgs = NULL)

Arguments

pkgs
Set of packages for which to report conflicts. If NULL, the default, will report conflicts for all loaded packages

Value

A named list of character vectors. The names are functions and the values are the packages where they appear. If there is only a single package listed, it means that an automated disambiguation has selected that function.

A user friendly print method displays the result as bulleted list.

Examples

conflict_scout()

[Package conflicted version 1.2.0 Index]